bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["dohodyaga <nnnzagora625@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
orjson = "^3.10.7"
aiogram = "^3.13.0"
httpx = "^0.27.2"
pydantic-settings = "^2.5.2"
pyjwt = "^2.9.0"
pyaml = "^24.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
